+++ test/bigtest/tests/sets/view-test.js
@@ -15,7 +15,7 @@ describe('Sets', () => {
 
   describe('View', () => {
     beforeEach(function () {
-      setsTest = this.server.create('sets');
+      setsTest = this.server.create('set');
 
       this.visit(`/settings/oai-pmh/sets/${setsTest.id}/view`);
     });